requirements.txt
bcrypt==3.2.0
cffi==1.15.0
click==8.0.3
colorama==0.4.4
dnspython==2.1.0
email-validator==1.1.3
Flask==2.0.2
Flask-Bcrypt==0.7.1
Flask-MySQLdb==0.2.0
Flask-WTF==1.0.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
mysqlclient==2.1.0
pycparser==2.21
python-dotenv==0.19.2
six==1.16.0
Werkzeug==2.0.2
WTForms==3.0.0